+++ inlang/source-code/end-to-end-tests/inlang-nextjs/app/[lng]/page.js
@@ -10,6 +10,7 @@ export default async function Page({ params: { lng } }) {
 			<br />
 			<Link href={`/${lng}/client-page`}>{t("translation:to-client-page")}</Link>
 			<Link href={`/${lng}/client-page`}>{t("translation:missingReference")}</Link>
+			<Link href={`/${lng}/client-page`}>{t("translation:missingReference")}</Link>
 			<Footer lng={lng} />
 		</>
 	)