+++ inlang/source-code/message-lint-rules/messageWithoutSource/src/messageWithoutSource.test.ts
@@ -52,4 +52,8 @@ test("should report if source message is missing", async () => {
 	expect(result.errors).toHaveLength(0)
 	expect(result.data).toHaveLength(1)
 	expect(result.data[0]!.languageTag).toBe("it")
+	expect(result.data[0]!.messageId).toBe(message1.id)
+	expect(
+		typeof result.data[0]!.body === "object" ? result.data[0]!.body.en : result.data[0]!.body,
+	).toContain(message1.id)
 })