+++ inlang/source-code/website/src/pages/editor/@host/@owner/@repository/components/Listheader.tsx
@@ -3,7 +3,8 @@ import { For, Show, createMemo } from "solid-js"
 import { showFilteredMessage } from "./../helper/showFilteredMessage.js"
 import { TourHintWrapper } from "./Notification/TourHintWrapper.jsx"
 import IconArrowLeft from "~icons/material-symbols/arrow-back-rounded"
-import type { InstalledMessageLintRule, MessageLintRule } from "@inlang/sdk"
+import { type InstalledMessageLintRule, type MessageLintRule } from "@inlang/sdk"
+import { navigate } from "vite-plugin-ssr/client/router"
 
 interface ListHeaderProps {
 	ids: string[]
@@ -23,6 +24,7 @@ export const ListHeader = (props: ListHeaderProps) => {
 		project,
 		setFilteredMessageLintRules,
 		filteredMessageLintRules,
+		filteredLanguageTags,
 		filteredId,
 		setFilteredId,
 		setTourStep,
@@ -33,8 +35,11 @@ export const ListHeader = (props: ListHeaderProps) => {
 		const summary: Record<MessageLintRule["id"], number> = {}
 		for (const report of project()?.query.messageLintReports.getAll() || []) {
 			if (
-				filteredMessageLintRules().length === 0 ||
-				filteredMessageLintRules().includes(report.ruleId)
+				((filteredMessageLintRules().length === 0 ||
+					filteredMessageLintRules().includes(report.ruleId)) &&
+					filteredLanguageTags().includes(report.languageTag) &&
+					filteredId() === "") ||
+				filteredId() === report.messageId
 			) {
 				summary[report.ruleId] = (summary[report.ruleId] || 0) + 1
 			}
@@ -145,6 +150,24 @@ export const ListHeader = (props: ListHeaderProps) => {
 						</Show>
 					)}
 				</For>
+				<Show when={project()?.installed.messageLintRules().length === 0}>
+					<sl-tooltip
+						prop:content={
+							"Install lint rules from the marketplace. They will help you write better translations."
+						}
+						prop:placement="bottom"
+						prop:trigger="hover"
+						style={{ "--show-delay": "1s" }}
+					>
+						<sl-button
+							prop:size="small"
+							// @ts-ignore
+							onClick={() => navigate("/marketplace")}
+						>
+							Install lint rules
+						</sl-button>
+					</sl-tooltip>
+				</Show>
 			</div>
 		</div>
 	)