+++ inlang/source-code/website/src/pages/marketplace/@id/index.page.server.tsx
@@ -3,14 +3,19 @@ import { convert } from "@inlang/markdown"
 import type { PageContext } from "#src/renderer/types.js"
 import type { PageProps } from "./index.page.jsx"
 import type { MarketplaceManifest } from "@inlang/marketplace-manifest"
+import { redirect } from "vite-plugin-ssr/abort"
 
 export async function onBeforeRender(pageContext: PageContext) {
 	const item = registry.find(
-		(item) => item.id === pageContext.routeParams.id,
+		(item) => item.id === pageContext.routeParams.id
 	) as MarketplaceManifest
 
-	const text = await(
-		// prettier-ignore
+	if (!item) {
+		console.error("Item not found")
+		throw redirect("/marketplace/404")
+	}
+
+	const text = await (
 		await fetch(typeof item.readme === "object" ? item.readme.en : item.readme)
 	).text()
 	const markdown = await convert(text)